nghttpx: Split up request class definition to dedicated files
